requirements.txt
confluent-kafka==2.7.0
Flask==2.3.2
python-dotenv==1.0.1
